+++ utils/graphql_queries.go
@@ -15,28 +15,6 @@ const (
             }
         }`
 
-	CommitCsQuery = `
-        query MyQuery($round: String!, $msgSender: String!) {
-            commitCs(where: {round: $round, msgSender: $msgSender}) {
-                blockTimestamp
-                commitVal
-            }
-        }`
-
-	RecoveredDataQuery = `
-        query MyQuery($round: String!) {
-            recovereds(orderBy: blockTimestamp, orderDirection: asc, where: {round: $round}) {
-                round
-                blockTimestamp
-                id
-                msgSender
-                omega
-                roundInfo {
-                    isRecovered
-                }
-            }
-        }`
-
 	CommitDataQuery = `
         query MyQuery($round: String!) {
             commits(where: {round: $round}){
@@ -56,15 +34,6 @@ const (
             round
         }
     }`
-
-	FulfillRandomnessDataQuery = `
-        query GetFulfillRandomness($round: String!) {
-            fulfillRandomnesses(where: {round: $round}) {
-                msgSender
-                blockTimestamp
-                success
-            }
-        }`
 
 	IsOperatorQuery = `
 		query MyQuery {
@@ -96,21 +65,6 @@ func GetRandomWordsRequestedRequest() *graphql.Request {
 	return graphql.NewRequest(RandomWordsRequestedQuery)
 }
 
-// GetCommitCsRequest returns a GraphQL request for fetching commitCs data.
-func GetCommitCsRequest(round, msgSender string) *graphql.Request {
-	req := graphql.NewRequest(CommitCsQuery)
-	req.Var("round", round)
-	req.Var("msgSender", msgSender)
-	return req
-}
-
-// GetRecoveredDataRequest returns a GraphQL request for fetching recovered data.
-func GetRecoveredDataRequest(round string) *graphql.Request {
-	req := graphql.NewRequest(RecoveredDataQuery)
-	req.Var("round", round)
-	return req
-}
-
 // GetCommitDataRequest returns a GraphQL request for fetching commit data.
 func GetCommitDataRequest(round string) *graphql.Request {
 	req := graphql.NewRequest(CommitDataQuery)
@@ -121,13 +75,6 @@ func GetCommitDataRequest(round string) *graphql.Request {
 // GetRevealDataRequest returns a GraphQL request for fetching commit data.
 func GetRevealDataRequest(round string) *graphql.Request {
 	req := graphql.NewRequest(RevealDataQuery)
-	req.Var("round", round)
-	return req
-}
-
-// GetFulfillRandomnessDataRequest returns a GraphQL request for fetching fulfill randomness data.
-func GetFulfillRandomnessDataRequest(round string) *graphql.Request {
-	req := graphql.NewRequest(FulfillRandomnessDataQuery)
 	req.Var("round", round)
 	return req
 }
